// new-app/src/entities/parameter/model/types.ts
export type EmployeeBase = {
  id: number;
  name: string;
};

export type ExtrusionDataRaw = {
  id: number;
  summary_id: number;
  counter_value: number;
  press_speed: number;
  blow_time: number;
  turning_machine_speed: number;
  annealing_furnace_temp: number;
  tube_cylindrical_section_length: number;
  membrane_thickness: number;
  tube_diameter: number;
  tube_cylindrical_section_thickness: number;
  tube_rigidity: number;
  tube_cutting_quality: boolean;
  tightness: boolean;
  external_thread_quality: boolean;
  tube_marking: boolean;
  employee_id: number;
  createdAt: Date;
  employee: EmployeeBase;
};

export type VarnishDataRaw = {
  id: number;
  summary_id: number;
  counter_value: number;
  varnish_machine_speed: number;
  total_air_pressure: number;
  feed_can_air_pressure: number;
  nozzle_regulator_air_pressure: number;
  cells_speed: number;
  injection_a_start_position: number;
  injection_b_start_position: number;
  injection_c_start_position: number;
  injection_d_start_position: number;
  injection_a_end_position: number;
  injection_b_end_position: number;
  injection_c_end_position: number;
  injection_d_end_position: number;
  tube_molding_start_position: number;
  tube_molding_end_position: number;
  polimerization_furnace_temp: number;
  internal_varnish_porosity: number;
  internal_sectional_view: boolean;
  aluminium_clearance_lack: boolean;
  unpainting_lack: boolean;
  employee_id: number;
  createdAt: Date;
  employee: EmployeeBase;
};

export type OffsetDataRaw = {
  id: number;
  summary_id: number;
  counter_value: number;
  printing_machine_speed: number;
  total_air_pressure: number;
  padding_furnace_temp: number;
  offset_furnace_temp: number;
  printer_motor: number;
  base_covers_holders_motor: number;
  base_covers_station_motor: number;
  imprint_quantity_printed_box_1: number | null;
  imprint_quantity_printed_box_2: number | null;
  imprint_quantity_printed_box_3: number | null;
  imprint_quantity_printed_box_4: number | null;
  imprint_quantity_printed_box_5: number | null;
  imprint_quantity_printed_box_6: number | null;
  ink_supply_time: number;
  design_match: boolean;
  tube_appearance: boolean;
  tube_edge_deformation_lack: boolean;
  aluminium_clearance_lack: boolean;
  drips_lack: boolean;
  employee_id: number;
  createdAt: Date;
  employee: EmployeeBase;
};

export type SealantDataRaw = {
  id: number;
  summary_id: number;
  counter_value: number;
  cap_machine_speed: number;
  total_air_pressure: number;
  holders_forward: number;
  holders_opening_left: number;
  holders_opening_right: number;
  holders_closing: number;
  injection_a_start: number;
  injection_b_start: number;
  injection_a_end: number;
  injection_b_end: number;
  injection_tube_orientation_start: number;
  injection_tube_orientation_end: number;
  is_cap_surface_smooth: boolean;
  latex_ring_padding: number;
  latex_ring_width: number;
  tube_rigidity: number;
  cap_unscrewing_torque: number;
  employee_id: number;
  createdAt: Date;
  employee: EmployeeBase;
};
// ...
export type ParameterSummaryData = {
  isActive: boolean;
  isFinished: boolean;
  productCode: string;
  productName: string;
  batchName: string;
  date: Date;
  shift: number;
  plan: number;
};

export type ExtrusionData = {
  summary: ParameterSummaryData;
  parameters: ExtrusionDataRaw;
  tresholds: Treshold;
  prev: number | null;
  next: number | null;
};

export type VarnishData = {
  summary: ParameterSummaryData;
  parameters: VarnishDataRaw;
  tresholds: Treshold;
  prev: number | null;
  next: number | null;
};

export type OffsetData = {
  summary: ParameterSummaryData;
  parameters: OffsetDataRaw;
  tresholds: Treshold;
  prev: number | null;
  next: number | null;
};

export type SealantData = {
  summary: ParameterSummaryData;
  parameters: SealantDataRaw;
  tresholds: Treshold;
  prev: number | null;
  next: number | null;
};

// Объединение для компонента ParameterDetails
export type AnyParameter =
  | { type: "extrusion"; data: ExtrusionData | null }
  | { type: "varnish"; data: VarnishData | null }
  | { type: "offset"; data: OffsetData | null }
  | { type: "sealant"; data: SealantData | null };